Extract restart-stable shared cells
